mix card map
