+++ src/03-shop/03-shop-card.js
@@ -1,0 +1,176 @@
+
+function ShopCard(){
+    return<>
+                        <div className="r-card-wrap">
+                  <div className="r-col">
+                    <div className="r-card-container">
+                        <div className="r-card-img-wrap">
+                          <img src='/03-shop-img/01cover.jpg' alt=""/>
+                        </div>
+                      <div className="r-card-body">
+                        <h2>惜食店鋪SEIZEE</h2>
+                        <div className="r-card-week-btn">
+                          <small>一</small>
+                          <small>二</small>
+                          <small>三</small>
+                          <small>四</small>
+                          <small>五</small>
+                          <small>六</small>
+                          <small>日</small>
+                        </div>
+                        <p>02-12345678</p>
+                        <p>營業時間:<span>9:00-21:00</span></p>
+                        <p>台北市大安區復興南路</p>
+                        
+                        <div className="r-card-button">
+                          <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                            <span>去逛逛</span>
+                          </a>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                  {/* <div className="r-col">
+                    <div className="r-card-container">
+                        <div className="r-card-img-wrap">
+                          <img src='/03-shop-img/01cover.jpg' alt=""/>
+                        </div>
+                      <div className="r-card-body">
+                        <h2>惜食店鋪SEIZEE</h2>
+                        <div className="r-card-week-btn">
+                          <small>一</small>
+                          <small>二</small>
+                          <small>三</small>
+                          <small>四</small>
+                          <small>五</small>
+                          <small>六</small>
+                          <small>日</small>
+                        </div>
+                        <p>02-12345678</p>
+                        <p>營業時間:<span>9:00-21:00</span></p>
+                        <p>台北市大安區復興南路一段390號哈哈哈哈哈哈哈</p>
+                        
+                        <div className="r-card-button">
+                          <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                            <span>去逛逛</span>
+                          </a>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="r-col">
+                    <div className="r-card-container">
+                        <div className="r-card-img-wrap">
+                          <img src='/03-shop-img/01cover.jpg' alt=""/>
+                        </div>
+                      <div className="r-card-body">
+                        <h2>惜食店鋪SEIZEE</h2>
+                        <div className="r-card-week-btn">
+                          <small>一</small>
+                          <small>二</small>
+                          <small>三</small>
+                          <small>四</small>
+                          <small>五</small>
+                          <small>六</small>
+                          <small>日</small>
+                        </div>
+                        <p>02-12345678</p>
+                        <p>營業時間:<span>9:00-21:00</span></p>
+                        <p>台北市大安區復興南路一段390號</p>
+                        
+                        <div className="r-card-button">
+                          <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                            <span>去逛逛</span>
+                          </a>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="r-col">
+                    <div className="r-card-container">
+                        <div className="r-card-img-wrap">
+                          <img src='/03-shop-img/01cover.jpg' alt=""/>
+                        </div>
+                      <div className="r-card-body">
+                        <h2>惜食店鋪SEIZEE</h2>
+                        <div className="r-card-week-btn">
+                          <small>一</small>
+                          <small>二</small>
+                          <small>三</small>
+                          <small>四</small>
+                          <small>五</small>
+                          <small>六</small>
+                          <small>日</small>
+                        </div>
+                        <p>02-12345678</p>
+                        <p>營業時間:<span>9:00-21:00</span></p>
+                        <p>台北市大安區復興南路一段390拉拉拉拉拉拉拉阿拉ㄌ</p>
+                        
+                        <div className="r-card-button">
+                          <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                            <span>去逛逛</span>
+                          </a>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="r-col">
+                    <div className="r-card-container">
+                        <div className="r-card-img-wrap">
+                          <img src='/03-shop-img/01cover.jpg' alt=""/>
+                        </div>
+                      <div className="r-card-body">
+                        <h2>惜食店鋪SEIZEE</h2>
+                        <div className="r-card-week-btn">
+                          <small>一</small>
+                          <small>二</small>
+                          <small>三</small>
+                          <small>四</small>
+                          <small>五</small>
+                          <small>六</small>
+                          <small>日</small>
+                        </div>
+                        <p>02-12345678</p>
+                        <p>營業時間:<span>9:00-21:00</span></p>
+                        <p>台北市大安區復</p>
+                        
+                        <div className="r-card-button">
+                          <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                            <span>去逛逛</span>
+                          </a>
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                  <div className="r-col">
+                    <div className="r-card-container">
+                        <div className="r-card-img-wrap">
+                          <img src='/03-shop-img/01cover.jpg' alt=""/>
+                        </div>
+                      <div className="r-card-body">
+                        <h2>惜食店鋪SEIZEE</h2>
+                        <div className="r-card-week-btn">
+                          <small>一</small>
+                          <small>二</small>
+                          <small>三</small>
+                          <small>四</small>
+                          <small>五</small>
+                          <small>六</small>
+                          <small>日</small>
+                        </div>
+                        <p>02-12345678</p>
+                        <p>營業時間:<span>9:00-21:00</span></p>
+                        <p>台北市大安區復興南路一段390號</p>
+                        
+                        <div className="r-card-button">
+                          <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                            <span>去逛逛</span>
+                          </a>
+                        </div>
+                      </div>
+                    </div>
+                  </div> */}
+                </div>
+    </>
+}
+export default ShopCard
