+++ src/03-shop/03-shop-m-card.js
@@ -1,0 +1,136 @@
+function ShopMcard() {
+    return (
+      <>
+
+        <div className="r-m-card-wrap">
+            <div className="r-m-col">
+            <div className="r-m-card-container">
+                <div class="r-m-card-img">
+                    <div className="r-m-card-img-wrap">
+                    <img src='/03-shop-img/01cover.jpg' alt=""/>
+                    </div>
+                    <p>營業中</p>
+                </div>
+                <div className="r-m-card-body">
+                <h2>惜食店鋪SEIZEE</h2>
+                <div className="r-m-card-week-btn">
+                    <small>一</small>
+                    <small>二</small>
+                    <small>三</small>
+                    <small>四</small>
+                    <small>五</small>
+                    <small>六</small>
+                    <small>日</small>
+                </div>
+                <p>02-12345678</p>
+                <p>營業時間:<span>9:00-21:00</span></p>
+                <p>台北市大安區復興南路一段390號</p>
+                
+                <div className="r-m-card-button">
+                    <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                    <span>去逛逛</span>
+                    </a>
+                </div>
+                </div>
+            </div>
+            </div>
+            <div className="r-m-col">
+            <div className="r-m-card-container">
+                <div class="r-m-card-img">
+                    <div className="r-m-card-img-wrap">
+                    <img src='/03-shop-img/01cover.jpg' alt=""/>
+                    </div>
+                    <p>營業中</p>
+                </div>
+                <div className="r-m-card-body">
+                <h2>惜食店鋪SEIZEE</h2>
+                <div className="r-m-card-week-btn">
+                    <small>一</small>
+                    <small>二</small>
+                    <small>三</small>
+                    <small>四</small>
+                    <small>五</small>
+                    <small>六</small>
+                    <small>日</small>
+                </div>
+                <p>02-12345678</p>
+                <p>營業時間:<span>9:00-21:00</span></p>
+                <p>台北市大安區復興南路一段390號</p>
+                
+                <div className="r-m-card-button">
+                    <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                    <span>去逛逛</span>
+                    </a>
+                </div>
+                </div>
+            </div>
+            </div>
+            <div className="r-m-col">
+            <div className="r-m-card-container">
+                <div class="r-m-card-img">
+                    <div className="r-m-card-img-wrap">
+                    <img src='/03-shop-img/01cover.jpg' alt=""/>
+                    </div>
+                    <p>營業中</p>
+                </div>
+                <div className="r-m-card-body">
+                <h2>惜食店鋪SEIZEE</h2>
+                <div className="r-m-card-week-btn">
+                    <small>一</small>
+                    <small>二</small>
+                    <small>三</small>
+                    <small>四</small>
+                    <small>五</small>
+                    <small>六</small>
+                    <small>日</small>
+                </div>
+                <p>02-12345678</p>
+                <p>營業時間:<span>9:00-21:00</span></p>
+                <p>台北市大安區復興南路一段390號</p>
+                
+                <div className="r-m-card-button">
+                    <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                    <span>去逛逛</span>
+                    </a>
+                </div>
+                </div>
+            </div>
+            </div>
+            <div className="r-m-col">
+            <div className="r-m-card-container">
+                <div class="r-m-card-img">
+                    <div className="r-m-card-img-wrap">
+                    <img src='/03-shop-img/01cover.jpg' alt=""/>
+                    </div>
+                    <p>營業中</p>
+                </div>
+                <div className="r-m-card-body">
+                <h2>惜食店鋪SEIZEE</h2>
+                <div className="r-m-card-week-btn">
+                    <small>一</small>
+                    <small>二</small>
+                    <small>三</small>
+                    <small>四</small>
+                    <small>五</small>
+                    <small>六</small>
+                    <small>日</small>
+                </div>
+                <p>02-12345678</p>
+                <p>營業時間:<span>9:00-21:00</span></p>
+                <p>台北市大安區復興南路一段390號</p>
+                
+                <div className="r-m-card-button">
+                    <a href="/#"><i className="fa-solid fa-caret-right"></i>
+                    <span>去逛逛</span>
+                    </a>
+                </div>
+                </div>
+            </div>
+            </div>
+
+        </div>
+
+      </>
+    )
+  }
+export default ShopMcard
