--- src/03-shop/DraggableMarker.js
@@ -1,45 +1,0 @@
-import { useState,useRef,useMemo,useCallback } from "react"
-import { MapContainer, TileLayer, useMap, Marker,Popup } from 'react-leaflet'
-
-const center = {
-    lat: 51.505,
-    lng: -0.09,
-  }
-  
-  function DraggableMarker() {
-    const [draggable, setDraggable] = useState(false)
-    const [position, setPosition] = useState(center)
-    const markerRef = useRef(null)
-    const eventHandlers = useMemo(
-      () => ({
-        dragend() {
-          const marker = markerRef.current
-          if (marker != null) {
-            setPosition(marker.getLatLng())
-          }
-        },
-      }),
-      [],
-    )
-    const toggleDraggable = useCallback(() => {
-      setDraggable((d) => !d)
-    }, [])
-  
-    return (
-      <Marker
-        draggable={draggable}
-        eventHandlers={eventHandlers}
-        position={position}
-        ref={markerRef}>
-        <Popup minWidth={90}>
-          <span onClick={toggleDraggable}>
-            {draggable
-              ? 'Marker is draggable'
-              : 'Click here to make marker draggable'}
-          </span>
-        </Popup>
-      </Marker>
-    )
-  }
-
-  export default DraggableMarker
